1.0

+++ client/src/components/users/utils/handleDeleteUser.js
@@ -1,11 +1,25 @@
 'use server';
+import { cookies } from 'next/headers';
 
-export async function deleteUser(userId, token) {
+export async function deleteUser(logic = false) {
     try {
-        const response = await fetch(`http://localhost:3000/api/users/${userId}`, {
+        const cookiesStore = cookies();
+        const token = cookiesStore.get('token')?.value;
+        const user = JSON.parse(cookiesStore.get('user')?.value || '{}');
+        const userId = user._id;
+
+        if (!token || !userId) {
+            throw new Error("Faltan credenciales para eliminar el usuario.");
+        }
+
+        // Construye la URL con el parámetro `logic` si es necesario
+        const url = `${process.env.API_URL}/users/${userId}${logic ? '?logic=true' : ''}`;
+
+        const response = await fetch(url, {
             method: 'DELETE',
             headers: {
                 'Authorization': `Bearer ${token}`,
+                'Content-Type': 'application/json',
             },
         });
